requirements.txt
fastapi
uvicorn
pandas
mysql-connector-python
python-dotenv
cryptography